Changes JSON button list to an array.
The button status list was originally an object with stringified
number keys because I thought it could be non-contiguous. However,
Logical button numbers assigned based on the physical layout with 0
at the top left and 48 at the bottom right will always be contiguous.

This converts those weird string-number-keyed objects into plain old
arrays of booleans. The array index is the logical button number, and
the value is the button state. This is a breaking API change.